Logs

+++ e-permit/src/main/java/epermit/services/KeyService.java
@@ -1,5 +1,6 @@
 package epermit.services;
 
+import java.util.List;
 import java.util.Optional;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -11,7 +12,9 @@ import epermit.repositories.AuthorityRepository;
 import epermit.repositories.KeyRepository;
 import epermit.utils.KeyUtil;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class KeyService {
@@ -22,16 +25,19 @@ public class KeyService {
 
     @Transactional
     public void create(String keyId) {
+        log.info("KeyService create started {}", keyId);
         Optional<Key> keyR = keyRepository.findOneByKeyId(keyId);
         if (keyR.isPresent()) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "KEYID_EXIST");
         }
         Key key = keyUtil.create(keyId);
+        log.info("KeyService create finished {}", key.getPublicJwk());
         keyRepository.save(key);
     }
 
     @Transactional
     public void enable(Integer id) {
+        log.info("KeyService enable started {}", id);
         Optional<Key> keyR = keyRepository.findById(id);
         if (!keyR.isPresent()) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "KEY_NOTFOUND");
@@ -47,9 +53,14 @@ public class KeyService {
 
     @Transactional
     public void delete(Integer id) {
+        log.info("KeyService delete started {}", id);
         Optional<Key> keyR = keyRepository.findById(id);
         if (!keyR.isPresent()) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "KEY_NOTFOUND");
+        }
+        List<Key> keys = keyRepository.findAllByEnabledTrue();
+        if(keys.size() == 1){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "THERE_IS_ONLY_ONE_KEY");
         }
         Key key = keyR.get();
         keyRepository.delete(key);
